feat: update the structure of directory.
