+++ index.js
@@ -2,7 +2,10 @@ import MonkeyMaster from "./main.js";
 import loadJsonFile from "https://deno.land/x/load_json_file@v1.0.0/mod.ts";
 import {
     logger
-} from "./log.js";
+} from "./util/log.js";
+import {
+    validateSkuids
+} from './util/validation.js';
 
 const CONFIG = await loadJsonFile("conf.json");
 
@@ -48,16 +51,4 @@ switch (mode) {
 
     default:
         break;
-}
-
-function validateSkuids(skuids) {
-    skuids.forEach((itemId) => {
-        if (!/^100/.test(itemId)) {
-            skuids = prompt(
-                "请输入100开头的抢购skuid,可以是多个，以逗号(,)分割",
-                "100016691566"
-            ).split(",");
-            validateSkuids(skuids);
-        }
-    });
 }
