+++ util/browser.js
